rename snapp.ts -> zkapp.ts
