+++ src/index.ts
@@ -5,7 +5,7 @@ export * from './lib/circuit_value';
 
 export * from './lib/int';
 export * as Mina from './lib/mina';
-export * from './lib/snapp';
+export * from './lib/zkapp';
 export * from './lib/optional';
 export * from './lib/proof_system';
 export * from './lib/party';
