+++ src/lib/zkapp.ts
